+++ repoze/what/plugins/pylonshq/protectors.py
@@ -22,6 +22,8 @@ All these utilities are also available in the
 
 """
 
+from decorator import decorator
+
 from pylons import request, response
 from pylons.controllers.util import abort
 from repoze.what.predicates import NotAuthorizedError
@@ -47,20 +49,13 @@ class _BaseProtectionDecorator(object):
         """
         self.predicate = predicate
         self.denial_handler = denial_handler or self.default_denial_handler
+    
 
 
 class ActionProtector(_BaseProtectionDecorator):
     """
     Function decorator to set predicate checkers in Pylons/TG2 controller
     actions.
-    
-    When authorization is denied, :func:`pylons.controllers.util.abort` will
-    be called with the 401 or 403 HTTP status code if the current user is
-    anonymous or authenticated, respectively.
-    
-    It's worth noting that when the status code for the response is 401,
-    that will trigger a :mod:`repoze.who` challenger (e.g., a login form will
-    be displayed).
     
     .. attribute:: default_denial_handler = None
     
@@ -70,33 +65,55 @@ class ActionProtector(_BaseProtectionDecorator):
     
     """
     
-    def __call__(self, func):
+    def __call__(self, action_):
         """
-        Return the decorator that will verify authorization when ``func`` is
-        run.
+        Return :meth:`wrap_action` as the decorator for ``action_``.
         
         """
+        return decorator(self.wrap_action, action_)
+    
+    def wrap_action(self, action_, *args, **kwargs):
+        """
+        Wrap the controller action ``action_``.
         
-        def check(*args, **kwargs):
-            try:
-                self.predicate.check_authorization(request.environ)
-            except NotAuthorizedError, e:
-                reason = unicode(e)
-                if request.environ.get('repoze.who.identity'):
-                    # The user is authenticated.
-                    code = 403
-                else:
-                    # The user is not authenticated.
-                    code = 401
-                if self.denial_handler:
-                    response.status = code
-                    return self.denial_handler(reason)
-                abort(code, comment=reason)
-            return func(*args, **kwargs)
-        # If we don't do the following, other decorators for ``func`` will get
-        # ignored:
-        check.__dict__.update(func.__dict__)
-        return check
+        :param action_: The controller action to be wrapped.
+        
+        ``args`` and ``kwargs`` are the positional and named arguments which
+        will be passed to ``action_`` when called.
+        
+        It will run ``action_`` if and only if authorization is granted (i.e.,
+        the predicate is met). Otherwise, it will set the HTTP status code
+        (to 401 if the user is anonymous or 403 if authenticated) then, if
+        defined, it will run the denial handler (if not, it will abort with
+        :func:`pylons.controllers.util.abort`).
+    
+        It's worth noting that when the status code for the response is 401,
+        that will trigger a :mod:`repoze.who` challenger (e.g., a login form 
+        will be displayed).
+        
+        .. note::
+        
+            If you want to override the default behavior when authorization is
+            denied (most likely), you should define just a denial handler. If
+            you want to override the whole wrapper (very unlikely), it's safe
+            to extend this class and override this method.
+        
+        """
+        try:
+            self.predicate.check_authorization(request.environ)
+        except NotAuthorizedError, e:
+            reason = unicode(e)
+            if request.environ.get('repoze.who.identity'):
+                # The user is authenticated.
+                code = 403
+            else:
+                # The user is not authenticated.
+                code = 401
+            if self.denial_handler:
+                response.status = code
+                return self.denial_handler(reason)
+            abort(code, comment=reason)
+        return action_(*args, **kwargs)
 
 
 class ControllerProtector(_BaseProtectionDecorator):
@@ -139,10 +156,13 @@ class ControllerProtector(_BaseProtectionDecorator):
             denial_handler = self.denial_handler
         else:
             denial_handler = getattr(cls, self.denial_handler)
+        
         if hasattr(cls, '__before__'):
-            old_before = cls.__before__
+            old_before = cls.__before__.im_func
         else:
             def old_before(*args, **kwargs): pass
+            old_before.__name__ = '__before__'
+        
         protector = self.protector(self.predicate, denial_handler)
         cls.__before__ = protector(old_before)